Add Toki Pona translation
Toki Pona ISO code is tok, but we register it as a dialect of Lojban (jbo), as the lingui depends on Unicode plural rules (https://www.unicode.org/cldr/charts/43/supplemental/language_plural_rules.html) which are not yet implemented for Toki Pona.
zh plural rules (none) are reused in i18n.ts as they are functionally identical.